feat: post 기능 추가 구현
- commet, board, user 단의 service를 import 하여 추가 기능 구현
- 뷰에서 호출한 게시판에 게시글이 작성되도록 구현
- 게시글 작성자이어여만 수정 및 삭제할 수 있도록 구현
- 게시글에서 댓글 기능 가능하도록 구현

+++ src/main/java/com/elice/boardproject/comment/controller/CommentController.java
@@ -41,7 +41,7 @@ public class CommentController {
     }
 
     @PostMapping("/comment/{comment_id}/edit") // 댓글 수정 요청
-    public String updateComment(@PathVariable int comment_id,
+    public String updateComment(@PathVariable Long comment_id,
                                 @ModelAttribute CommentDTO commentDTO,
                                 RedirectAttributes redirectAttributes,
                                 HttpSession session) {
@@ -62,7 +62,7 @@ public class CommentController {
     }
 
     @DeleteMapping("/comment/{comment_id}/delete") //댓글 삭제 요청
-    public String deleteComment(@PathVariable int comment_id,
+    public String deleteComment(@PathVariable Long comment_id,
                                 RedirectAttributes redirectAttributes,
                                 HttpSession session) {
